+++ crates/dxm-init/src/vcs.rs
@@ -5,27 +5,12 @@ use std::{error::Error, fmt::Display, path::Path, str::FromStr};
 use dxm_manifest::Manifest;
 use git2::Repository;
 
-const GITIGNORE_NAME: &str = ".gitignore";
+pub const GITIGNORE_NAME: &str = ".gitignore";
+pub const TEMPLATE_EXTENSION: &str = "example";
 
-const ROOT_GITIGNORE: &str = "\
-# FXServer
-/artifact/
-
-# txAdmin
-/txData/
-";
-
-const DATA_GITIGNORE: &str = "\
-# Cache
-/cache/
-
-# KVP
-/db/
-
-# Miscellaneous
-/.replxx_history
-/imgui.ini
-";
+const GIT_README: &str = include_str!("../templates/git/README.md");
+const ROOT_GITIGNORE: &str = include_str!("../templates/git/root.gitignore");
+const DATA_GITIGNORE: &str = include_str!("../templates/git/data.gitignore");
 
 /// The possible version control systems to use in servers.
 #[derive(Default, Debug, PartialEq, Eq, Clone)]
@@ -78,13 +63,26 @@ impl VcsOption {
             VcsOption::Git => {
                 Repository::init(path)?;
 
+                fs_err::write(crate::README_NAME, GIT_README)?;
                 fs_err::write(path.join(GITIGNORE_NAME), ROOT_GITIGNORE)?;
                 fs_err::write(data_path.join(GITIGNORE_NAME), DATA_GITIGNORE)?;
+
+                create_template(data_path.join(crate::ENV_CFG_NAME))?;
+                create_template(data_path.join(crate::SECRETS_CFG_NAME))?;
 
                 Ok(())
             }
         }
     }
+}
+
+fn create_template<P>(path: P) -> std::io::Result<u64>
+where
+    P: AsRef<Path>,
+{
+    let path = path.as_ref();
+
+    fs_err::copy(path, path.with_added_extension(TEMPLATE_EXTENSION))
 }
 
 #[cfg(test)]